requirements.txt
gevent==0.13.7
py-bcrypt==0.2
